Retrieve bird ids in one query
Removes n+1 queries when gathering descendants and their bird_ids
Similar to the changes made in the node model using WITH RECURSIVE.
Also joins the birds table to get the bird ids in the same query.

Like the previous implementation for the birds endpoint, it does not
raise an exception if one of the node ids is in a circular reference but
the UNION without the ALL accounts for and guards against that causing
an infinite loop.