Disable hashjoin in postgres
Something was causing the query planner to mis-estimate the cost of the
different join stragegies. It was choosing hashjoin which does not use
indexes over mergejoin which does.  Turning off the hashjoin strategy
sped up the joins of the WITH RECURSIVE loops.

If you don't want to turn it off across the database you could
temporarily disable it before doing the query and then turning it back
on again after.